Add files via upload
This project allows the user to select either a book or a short story from the New York Times Books API and Top Stories API respectively. The WHAT SHOULD I READ NEXT? website provides the general information needed to know about the book or story picked. The selection process is randomized, both for the book picking and story picking for the user.  If the user decided to read a book, the title, author, book ranking number, short description, and ISBN number will be provided along with the book cover displayed image. If the user decided to read a story from the top stories list, the title, author, genre section, abstract of the story, and the URL link will be provided along with an image from the story displayed.